chore(feed): move module to data and domain layer
