+++ app/src/androidTest/java/com/example/alpacamusclemaintenance/di/TestAppModule.kt
@@ -1,9 +1,7 @@
 package com.example.alpacamusclemaintenance.di
 
-import com.example.alpacamusclemaintenance.api.QiitaApi
 import com.example.alpacamusclemaintenance.db.AppDatabase
 import com.example.alpacamusclemaintenance.db.dao.PushUpDao
-import com.example.alpacamusclemaintenance.repository.FeedRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -19,10 +17,4 @@ class TestAppModule {
 
     @Provides
     fun providePushUpDao(): PushUpDao = mock(PushUpDao::class.java)
-
-    @Provides
-    fun provideFeedRepository(): FeedRepository = mock(FeedRepository::class.java)
-
-    @Provides
-    fun provideQiitaService(): QiitaApi = mock(QiitaApi::class.java)
 }
