--- app/src/main/java/com/example/alpacamusclemaintenance/vo/FeedSearchResult.kt
@@ -1,6 +1,0 @@
-package com.example.alpacamusclemaintenance.vo
-
-sealed class FeedSearchResult {
-    data class Success(val data: List<Feed>) : FeedSearchResult()
-    data class Error(val error: Exception) : FeedSearchResult()
-}
