+++ build.gradle
@@ -21,6 +21,7 @@ buildscript {
 
 ext {
     minSdkVersion = 26
+    buildToolsVersion = "30.0.3"
     targetSdkVersion = 30
     versionCode = 1
     versionName = "1.0"
