+++ data/src/main/java/com/example/alpacamusclemaintenance/data/feed/FeedPagingSource.kt
@@ -1,8 +1,7 @@
-package com.example.alpacamusclemaintenance.repository
+package com.example.alpacamusclemaintenance.data.feed
 
 import androidx.paging.PagingSource
-import com.example.alpacamusclemaintenance.api.QiitaApi
-import com.example.alpacamusclemaintenance.vo.Feed
+import com.example.alpacamusclemaintenance.domain.feed.Feed
 
 private const val QIITA_STARTING_PAGE_INDEX = 1
