+++ domain/src/main/java/com/example/alpacamusclemaintenance/domain/feed/FeedModel.kt
@@ -1,4 +1,4 @@
-package com.example.alpacamusclemaintenance.vo
+package com.example.alpacamusclemaintenance.domain.feed
 
 import java.util.Date
